test: setup online functions testing, organized test structure
