+++ functions/.env.EXAMPLE
@@ -1,0 +1,6 @@
+FIREBASE_DATABASE_URL=
+FIREBASE_STORAGE_BUCKET=
+FIREBASE_PROJECT_ID=
+ALGOLIA_KEY=
+ALGOLIA_ID=
+ALGOLIA_REQUESTS_INDEX=
