+++ functions/package.json
@@ -42,6 +42,7 @@
     "algoliasearch": "^4.1.0",
     "class-validator": "^0.11.1",
     "convict": "^5.2.0",
+    "dotenv": "^8.2.0",
     "firebase": "^7.14.0",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.8.0",
