+++ functions/tests/index.test.ts
@@ -1,135 +1,27 @@
-import { validateOrReject } from 'class-validator';
+import * as dotenv from 'dotenv';
+import * as Test from 'firebase-functions-test';
 
-import { ApplicationPreference, User } from '../src/models/users';
+dotenv.config();
 
-describe('user validation', () => {
-  it('fails when you pass it negative average ', () => {
-    const user = User.factory({
-      cavRatingsReceived: 0,
-      pinRatingsReceived: 0,
-      cavQuestionnaireRef: null,
-      pinQuestionnaireRef: null,
-      averageRating: -1,
-      casesCompleted: 0,
-      requestsMade: 0,
-      username: 'TestUser',
-      displayName: null,
-      displayPicture: null,
-      applicationPreference: ApplicationPreference.pin,
-    });
+const firebaseConfig = {
+  databaseURL: process.env.FIREBASE_DATABASE_URL,
+  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+  projectId: process.env.FIREBASE_PROJECT_ID,
+};
+const pathToServiceAccountKey = `${__dirname}/../serviceAccountKey-dev.json`;
+const firebaseFunctionsTest = Test(firebaseConfig, pathToServiceAccountKey);
+firebaseFunctionsTest.mockConfig({
+  algolia: {
+    id: process.env.ALGOLIA_ID,
+    key: process.env.ALGOLIA_KEY,
+    requests_index: process.env.ALGOLIA_REQUESTS_INDEX, // eslint-disable-line @typescript-eslint/camelcase
+  },
+});
 
-    return validateOrReject(user)
-      .then(() => {
-        throw Error('Validations should have failed, but they passed');
-      })
-      .catch(errors => {
-        expect(errors).toBeTruthy();
-      });
-  });
-
-  it('fails when you pass it negative cav ratings received ', () => {
-    const user = User.factory({
-      cavRatingsReceived: -1,
-      pinRatingsReceived: 0,
-      cavQuestionnaireRef: null,
-      pinQuestionnaireRef: null,
-      averageRating: 0,
-      casesCompleted: 0,
-      requestsMade: 0,
-      username: 'TestUser',
-      displayName: null,
-      displayPicture: null,
-      applicationPreference: ApplicationPreference.pin,
-    });
-
-    return validateOrReject(user)
-      .then(() => {
-        throw Error('Validations should have failed, but they passed');
-      })
-      .catch(errors => {
-        expect(errors).toBeTruthy();
-      });
-  });
-
-  it('fails when you pass it negative pin ratings received ', () => {
-    const user = User.factory({
-      cavRatingsReceived: 1,
-      pinRatingsReceived: -1,
-      cavQuestionnaireRef: null,
-      pinQuestionnaireRef: null,
-      averageRating: 0,
-      casesCompleted: 0,
-      requestsMade: 0,
-      username: 'TestUser',
-      displayName: null,
-      displayPicture: null,
-      applicationPreference: ApplicationPreference.pin,
-    });
-
-    return validateOrReject(user)
-      .then(() => {
-        throw Error('Validations should have failed, but they passed');
-      })
-      .catch(errors => {
-        expect(errors).toBeTruthy();
-      });
-  });
-
-  it('fails when you pass it negative cases completed', () => {
-    const user = User.factory({
-      cavRatingsReceived: 1,
-      pinRatingsReceived: 1,
-      cavQuestionnaireRef: null,
-      pinQuestionnaireRef: null,
-      averageRating: 0,
-      casesCompleted: -1,
-      requestsMade: 0,
-      username: 'TestUser',
-      displayName: null,
-      displayPicture: null,
-      applicationPreference: ApplicationPreference.pin,
-    });
-
-    return validateOrReject(user)
-      .then(() => {
-        throw Error('Validations should have failed, but they passed');
-      })
-      .catch(errors => {
-        expect(errors).toBeTruthy();
-      });
-  });
-
-  it('fails when you pass it negative requests made', () => {
-    const user = User.factory({
-      cavRatingsReceived: 1,
-      pinRatingsReceived: 1,
-      cavQuestionnaireRef: null,
-      pinQuestionnaireRef: null,
-      averageRating: 0,
-      casesCompleted: 0,
-      requestsMade: -1,
-      username: 'TestUser',
-      displayName: null,
-      displayPicture: null,
-      applicationPreference: ApplicationPreference.pin,
-    });
-
-    return validateOrReject(user)
-      .then(() => {
-        throw Error('Validations should have failed, but they passed');
-      })
-      .catch(errors => {
-        expect(errors).toBeTruthy();
-      });
-  });
-
-  it('succeeds when you pass it minimum info ', () => {
-    const user = User.factory({
-      username: 'test_user',
-    });
-
-    return validateOrReject(user).then(() => {
-      expect(true).toBeTruthy();
-    });
+describe('Jest test suite for functions', () => {
+  it('tests operational', () => {
+    expect(1 + 1).toBe(2);
   });
 });
+
+export { firebaseFunctionsTest };
